+++ src/error.rs
@@ -21,9 +21,18 @@ pub enum InquireError {
     #[error("IO error: {0}")]
     IO(#[from] io::Error),
 
-    /// The user canceled the operation by pressing Ctrl+C or ESC.
+    /// The user canceled the operation by pressing ESC.
     #[error("Operation was canceled by the user")]
     OperationCanceled,
+
+    /// The operation was interrupted by the user after they
+    /// pressed Ctrl+C.
+    ///
+    /// This error will be returned only when using `crossterm`
+    /// or `termion` as the terminal back-end. If using `console`,
+    /// pressing Ctrl+C will trigger SIGINT.
+    #[error("Operation was interrupted by the user")]
+    OperationInterrupted,
 }
 
 /// Result type where errors are of type [InquireError](crate::error::InquireError)