+++ src/prompts/mod.rs
@@ -1,18 +1,23 @@
+#[macro_use]
+pub(in crate) mod prompt_common;
+
 mod confirm;
 mod custom_type;
 #[cfg(feature = "date")]
 mod dateselect;
+#[cfg(feature = "editor")]
+mod editor;
 mod multiselect;
 mod password;
 mod select;
 mod text;
 
 pub use confirm::Confirm;
-
+pub use custom_type::CustomType;
 #[cfg(feature = "date")]
 pub use dateselect::DateSelect;
-
-pub use custom_type::CustomType;
+#[cfg(feature = "editor")]
+pub use editor::Editor;
 pub use multiselect::MultiSelect;
 pub use password::{Password, PasswordDisplayMode};
 pub use select::Select;